requirements.txt
astroid==2.3.3
attrs==19.3.0
bravado==10.5.0
bravado-core==5.16.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
cycler==0.10.0
docker==3.7.3
docker-pycreds==0.4.0
Flask==0.12.5
future==0.18.2
gitdb==4.0.2
GitPython==3.1.0
GPUtil==1.4.0
gym==0.10.11
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jsonmerge==1.5.2
jsonpointer==2.0
jsonref==0.2
jsonschema==2.6.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
monotonic==1.5
msgpack==1.0.0
msgpack-python==0.5.6
neptune-client==0.4.105
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.1
Pillow==5.4.1
pommerman==0.2.0
psutil==5.7.0
py3nvml==0.2.5
pyglet==1.5.0
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.7
python-cli-ui==0.7.5
python-dateutil==2.8.1
python-rapidjson==0.6.3
pytz==2019.3
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scipy==1.4.1
seaborn==0.10.0
simplejson==3.17.0
six==1.14.0
smmap==3.0.1
strict-rfc3339==0.7
swagger-spec-validator==2.5.0
tabulate==0.8.6
torch==1.4.0
typing-extensions==3.7.4.1
Unidecode==1.1.1
urllib3==1.25.8
webcolors==1.11.1
websocket-client==0.53.0
websockets==6.0
Werkzeug==0.16.1
wrapt==1.11.2
xmltodict==0.12.0
